requirements.txt
discord.py
yfinance
pandas
scikit-learn
joblib
python-dotenv
